Move `resolver` outside of `tunnel_state_machine`.
